Split unix.rs into separate files
