+++ src/x11.rs
@@ -149,168 +149,3 @@ impl XcbWindowHandle {
         }
     }
 }
-
-/// Raw display handle for Wayland.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct WaylandDisplayHandle {
-    /// A pointer to a `wl_display`.
-    pub display: NonNull<c_void>,
-}
-
-impl WaylandDisplayHandle {
-    /// Create a new display handle.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::ffi::c_void;
-    /// # use core::ptr::NonNull;
-    /// # use raw_window_handle::WaylandDisplayHandle;
-    /// #
-    /// let display: NonNull<c_void>;
-    /// # display = NonNull::from(&()).cast();
-    /// let handle = WaylandDisplayHandle::new(display);
-    /// ```
-    pub fn new(display: NonNull<c_void>) -> Self {
-        Self { display }
-    }
-}
-
-/// Raw window handle for Wayland.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct WaylandWindowHandle {
-    /// A pointer to a `wl_surface`.
-    pub surface: NonNull<c_void>,
-}
-
-impl WaylandWindowHandle {
-    /// Create a new handle to a surface.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::ffi::c_void;
-    /// # use core::ptr::NonNull;
-    /// # use raw_window_handle::WaylandWindowHandle;
-    /// #
-    /// let surface: NonNull<c_void>;
-    /// # surface = NonNull::from(&()).cast();
-    /// let handle = WaylandWindowHandle::new(surface);
-    /// ```
-    pub fn new(surface: NonNull<c_void>) -> Self {
-        Self { surface }
-    }
-}
-
-/// Raw display handle for the Linux Kernel Mode Set/Direct Rendering Manager.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct DrmDisplayHandle {
-    /// The drm file descriptor.
-    // TODO: Use `std::os::fd::RawFd`?
-    pub fd: i32,
-}
-
-impl DrmDisplayHandle {
-    /// Create a new handle to a file descriptor.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use raw_window_handle::DrmDisplayHandle;
-    /// #
-    /// let fd: i32;
-    /// # fd = 0;
-    /// let handle = DrmDisplayHandle::new(fd);
-    /// ```
-    pub fn new(fd: i32) -> Self {
-        Self { fd }
-    }
-}
-
-/// Raw window handle for the Linux Kernel Mode Set/Direct Rendering Manager.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct DrmWindowHandle {
-    /// The primary drm plane handle.
-    pub plane: u32,
-}
-
-impl DrmWindowHandle {
-    /// Create a new handle to a plane.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use raw_window_handle::DrmWindowHandle;
-    /// #
-    /// let plane: u32;
-    /// # plane = 0;
-    /// let handle = DrmWindowHandle::new(plane);
-    /// ```
-    pub fn new(plane: u32) -> Self {
-        Self { plane }
-    }
-}
-
-/// Raw display handle for the Linux Generic Buffer Manager.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct GbmDisplayHandle {
-    /// The gbm device.
-    pub gbm_device: NonNull<c_void>,
-}
-
-impl GbmDisplayHandle {
-    /// Create a new handle to a device.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::ffi::c_void;
-    /// # use core::ptr::NonNull;
-    /// # use raw_window_handle::GbmDisplayHandle;
-    /// #
-    /// let ptr: NonNull<c_void>;
-    /// # ptr = NonNull::from(&()).cast();
-    /// let handle = GbmDisplayHandle::new(ptr);
-    /// ```
-    pub fn new(gbm_device: NonNull<c_void>) -> Self {
-        Self { gbm_device }
-    }
-}
-
-/// Raw window handle for the Linux Generic Buffer Manager.
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct GbmWindowHandle {
-    /// The gbm surface.
-    pub gbm_surface: NonNull<c_void>,
-}
-
-impl GbmWindowHandle {
-    /// Create a new handle to a surface.
-    ///
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::ffi::c_void;
-    /// # use core::ptr::NonNull;
-    /// # use raw_window_handle::GbmWindowHandle;
-    /// #
-    /// let ptr: NonNull<c_void>;
-    /// # ptr = NonNull::from(&()).cast();
-    /// let handle = GbmWindowHandle::new(ptr);
-    /// ```
-    pub fn new(gbm_surface: NonNull<c_void>) -> Self {
-        Self { gbm_surface }
-    }
-}
